Use MUnit in CodecRouterSuite
